requirements.txt
streamlit>=1.30.0
huggingface_hub>=0.20.0
gspread>=6.0.0
google-auth>=2.25.0
folium>=0.15.0
streamlit-folium>=0.18.0
Pillow>=10.0.0
